Add Location in manageContent fix bug function nvl

+++ administrator/mod_category/content_action.php
@@ -46,20 +46,19 @@ if (isset($_GET['delete'])) {
 	$sql = "UPDATE trn_content_detail SET  " . implode(",", $update) . " WHERE CONTENT_ID =" . $id;
 
 	mysql_query($sql, $conn);
-	
+
 	if (count($_POST['video_delete_gallery']) > 0) {
 		foreach ($_POST['video_delete_gallery'] as $k => $file) {
-			$file = explode('|@|',$file);
+			$file = explode('|@|', $file);
 
-			if($file[0] == 'upload'){
+			if ($file[0] == 'upload') {
 				del_video_file($file[2]);
 			}
 
-			mysql_query('DELETE FROM trn_content_picture WHERE PIC_ID = '.intval($file[1]), $conn) or die($sql);
+			mysql_query('DELETE FROM trn_content_picture WHERE PIC_ID = ' . intval($file[1]), $conn) or die($sql);
 		}
 		mysql_query('OPTIMIZE TABLE trn_content_picture', $conn) or die("");
 	}
-	
 
 }
 
@@ -106,6 +105,10 @@ if (isset($_GET['add'])) {
 	$insert['EVENT_START_DATE'] = "'" . ConvertDateToDB($_POST['txtStartDate']) . "'";
 	$insert['EVENT_END_DATE'] = "'" . ConvertDateToDB($_POST['txtEndDate']) . "'";
 
+	$insert['PLACE_DESC_LOC'] = "'" . nvl($_POST['txtPlaceLoc'], "") . "'";
+	$insert['PLACE_DESC_ENG'] = "'" . nvl($_POST['txtPlaceEng'], "") . "'";
+	$insert['LAT'] = "'" . nvl($_POST['txtLat'], "") . "'";
+	$insert['LON'] = "'" . nvl($_POST['txtLon'], "") . "'";
 	$sql = "INSERT INTO  trn_content_detail (" . implode(",", array_keys($insert)) . ") VALUES (" . implode(",", array_values($insert)) . ")";
 
 	mysql_query($sql, $conn) or die($sql);
@@ -113,7 +116,7 @@ if (isset($_GET['add'])) {
 
 	if (count($_POST['photo_file']) > 0) {
 		$index = 1;
-		 
+
 		foreach ($_POST['photo_file'] as $k => $file) {
 			$filename = admin_move_image_upload_dir('content_' . $CID, end(explode('/', $file)), 1000, '', false, 150, 150);
 
@@ -129,62 +132,62 @@ if (isset($_GET['add'])) {
 			mysql_query($sql, $conn) or die($sql);
 		}
 	}
-	
+
 	if (count($_POST['video_video']) > 0) {
 		$index = 1;
 		foreach ($_POST['video_video'] as $k => $file) {
 
-			$file = explode('|@|',$file);
+			$file = explode('|@|', $file);
 
-			if($file[0] == 'upload'){
+			if ($file[0] == 'upload') {
 				$IMG_TYPE = 2;
-				$file[1] = move_video_file($file[1],'content_'.$CID);
-			}else if($file[0] == 'embed'){
+				$file[1] = move_video_file($file[1], 'content_' . $CID);
+			} else if ($file[0] == 'embed') {
 				$IMG_TYPE = 3;
-			}else if($file[0] == 'link'){
+			} else if ($file[0] == 'link') {
 				$IMG_TYPE = 4;
 			}
-
 
 			unset($insert);
 			$insert['CONTENT_ID'] = $retrunID;
 			$insert['IMG_TYPE'] = $IMG_TYPE;
 			$insert['IMG_PATH'] = "'" . $file[1] . "'";
-			$insert['CAT_ID'] = $CID ;
+			$insert['CAT_ID'] = $CID;
 			$insert['ORDER_ID'] = "'" . $index++ . "'";
 			$insert['IMG_NAME'] = "'" . $file[2] . "'";
-			$insert['DIV_NAME'] = "'video'"; //ตั้งตาม name
+			$insert['DIV_NAME'] = "'video'";
+			//ตั้งตาม name
 
 			$sql = "INSERT INTO trn_content_picture (" . implode(",", array_keys($insert)) . ") VALUES (" . implode(",", array_values($insert)) . ")";
 			mysql_query($sql, $conn) or die($sql);
 
 		}
 	}
-	
+
 	if (count($_POST['video_other']) > 0) {
 		$index = 1;
 		foreach ($_POST['video_other'] as $k => $file) {
 
-			$file = explode('|@|',$file);
+			$file = explode('|@|', $file);
 
-			if($file[0] == 'upload'){
+			if ($file[0] == 'upload') {
 				$IMG_TYPE = 2;
-				$file[1] = move_video_file($file[1],'content_'.$CID);
-			}else if($file[0] == 'embed'){
+				$file[1] = move_video_file($file[1], 'content_' . $CID);
+			} else if ($file[0] == 'embed') {
 				$IMG_TYPE = 3;
-			}else if($file[0] == 'link'){
+			} else if ($file[0] == 'link') {
 				$IMG_TYPE = 4;
 			}
-
 
 			unset($insert);
 			$insert['CONTENT_ID'] = $retrunID;
 			$insert['IMG_TYPE'] = $IMG_TYPE;
 			$insert['IMG_PATH'] = "'" . $file[1] . "'";
-			$insert['CAT_ID'] = $CID ;
+			$insert['CAT_ID'] = $CID;
 			$insert['ORDER_ID'] = "'" . $index++ . "'";
 			$insert['IMG_NAME'] = "'" . $file[2] . "'";
-			$insert['DIV_NAME'] = "'other'"; //ตั้งตาม name
+			$insert['DIV_NAME'] = "'other'";
+			//ตั้งตาม name
 
 			$sql = "INSERT INTO trn_content_picture (" . implode(",", array_keys($insert)) . ") VALUES (" . implode(",", array_values($insert)) . ")";
 			mysql_query($sql, $conn) or die($sql);
@@ -192,6 +195,37 @@ if (isset($_GET['add'])) {
 		}
 	}
 
+	//noppol.von 18Oct2015 add Audio File
+	if (count($_POST['video_sound']) > 0) {
+		$index = 1;
+		foreach ($_POST['video_sound'] as $k => $file) {
+
+			$file = explode('|@|', $file);
+
+			if ($file[0] == 'upload') {
+				$IMG_TYPE = 2;
+				$file[1] = move_video_file($file[1], 'content_' . $CID);
+			} else if ($file[0] == 'embed') {
+				$IMG_TYPE = 3;
+			} else if ($file[0] == 'link') {
+				$IMG_TYPE = 4;
+			}
+
+			unset($insert);
+			$insert['CONTENT_ID'] = $retrunID;
+			$insert['IMG_TYPE'] = $IMG_TYPE;
+			$insert['IMG_PATH'] = "'" . $file[1] . "'";
+			$insert['CAT_ID'] = $CID;
+			$insert['ORDER_ID'] = "'" . $index++ . "'";
+			$insert['IMG_NAME'] = "'" . $file[2] . "'";
+			$insert['DIV_NAME'] = "'sound'";
+			//ตั้งตาม name
+
+			$sql = "INSERT INTO trn_content_picture (" . implode(",", array_keys($insert)) . ") VALUES (" . implode(",", array_values($insert)) . ")";
+			mysql_query($sql, $conn) or die($sql);
+
+		}
+	}
 	header('Location: ' . $returnPage);
 	//echo $scriptReturnPath ;
 }
@@ -225,6 +259,11 @@ if (isset($_GET['edit'])) {
 	$update[] = "EVENT_START_DATE = '" . ConvertDateToDB($_POST['txtStartDate']) . "'";
 	$update[] = "EVENT_END_DATE = '" . ConvertDateToDB($_POST['txtEndDate']) . "'";
 
+	$update[] = "PLACE_DESC_LOC = '" . nvl($_POST['txtPlaceLoc'], "") . "'";
+	$update[] = "PLACE_DESC_ENG = '" . nvl($_POST['txtPlaceEng'], "") . "'";
+	$update[] = "LAT = '" . nvl($_POST['txtLat'], "") . "'";
+	$update[] = "LON  = '" . nvl($_POST['txtLon'], "") . "'";
+
 	$sql = "UPDATE trn_content_detail SET  " . implode(",", $update) . " WHERE CONTENT_ID = " . $conid;
 	mysql_query($sql, $conn);
 
@@ -235,7 +274,6 @@ if (isset($_GET['edit'])) {
 		$row_max = mysql_fetch_array($query_max);
 		$max = $row_max['MAX_ORDER'];
 		$max++;
-
 
 		foreach ($_POST['photo_file'] as $k => $file) {
 			$filename = admin_move_image_upload_dir('content_' . $CID, end(explode('/', $file)), 1000, '', false, 150, 150);
@@ -252,14 +290,13 @@ if (isset($_GET['edit'])) {
 			mysql_query($sql, $conn) or die($sql);
 		}
 	}
-	
-	
+
 	if (count($_POST['video_other']) > 0) {
 		$CONTENT_ID = intval($conid);
-		$CAT_ID		= intval($CID);
-		$DIV_NAME	= 'other';
+		$CAT_ID = intval($CID);
+		$DIV_NAME = 'other';
 
-		$sql = "SELECT ORDER_ID FROM trn_content_picture WHERE CONTENT_ID = ".$CONTENT_ID." AND CAT_ID = ".$CAT_ID." AND DIV_NAME =  '".$DIV_NAME."' ORDER BY ORDER_ID DESC LIMIT 0 , 1";
+		$sql = "SELECT ORDER_ID FROM trn_content_picture WHERE CONTENT_ID = " . $CONTENT_ID . " AND CAT_ID = " . $CAT_ID . " AND DIV_NAME =  '" . $DIV_NAME . "' ORDER BY ORDER_ID DESC LIMIT 0 , 1";
 		$query = mysql_query($sql, $conn) or die($sql);
 		$row = mysql_fetch_array($query);
 		$index = $row['ORDER_ID'];
@@ -267,52 +304,109 @@ if (isset($_GET['edit'])) {
 
 		foreach ($_POST['video_other'] as $k => $file) {
 
-			$file = explode('|@|',$file);
+			$file = explode('|@|', $file);
 
-			if($file[0] == 'upload'){
+			if ($file[0] == 'upload') {
 				$IMG_TYPE = 2;
-				$file[1] = move_video_file($file[1],'content_' . $CID);
-			}else if($file[0] == 'embed'){
+				$file[1] = move_video_file($file[1], 'content_' . $CID);
+			} else if ($file[0] == 'embed') {
 				$IMG_TYPE = 3;
-			}else if($file[0] == 'link'){
+			} else if ($file[0] == 'link') {
 				$IMG_TYPE = 4;
 			}
 
-
 			unset($insert);
-			$insert['CONTENT_ID'] = $CONTENT_ID; /*retrunID*/
+			$insert['CONTENT_ID'] = $CONTENT_ID;
+			/*retrunID*/
 			$insert['IMG_TYPE'] = $IMG_TYPE;
 			$insert['IMG_PATH'] = "'" . $file[1] . "'";
-			$insert['CAT_ID'] = $CAT_ID; /* cat ID*/
+			$insert['CAT_ID'] = $CAT_ID;
+			/* cat ID*/
 			$insert['ORDER_ID'] = "'" . $index++ . "'";
 			$insert['IMG_NAME'] = "'" . $file[2] . "'";
-			$insert['DIV_NAME'] = "'".$DIV_NAME."'"; //ตั้งตาม name
+			$insert['DIV_NAME'] = "'" . $DIV_NAME . "'";
+			//ตั้งตาม name
 
 			$sql = "INSERT INTO trn_content_picture (" . implode(",", array_keys($insert)) . ") VALUES (" . implode(",", array_values($insert)) . ")";
 			mysql_query($sql, $conn) or die($sql);
 
 		}
 	}
-	//video action delete
-	if (count($_POST['video_delete_other']) > 0) {
-		foreach ($_POST['video_delete_other'] as $k => $file) {
-			$file = explode('|@|',$file);
 
-			if($file[0] == 'upload'){
+if (count($_POST['video_sound']) > 0) {
+		$CONTENT_ID = intval($conid);
+		$CAT_ID = intval($CID);
+		$DIV_NAME = 'sound';
+
+		$sql = "SELECT ORDER_ID FROM trn_content_picture WHERE CONTENT_ID = " . $CONTENT_ID . " AND CAT_ID = " . $CAT_ID . " AND DIV_NAME =  '" . $DIV_NAME . "' ORDER BY ORDER_ID DESC LIMIT 0 , 1";
+		$query = mysql_query($sql, $conn) or die($sql);
+		$row = mysql_fetch_array($query);
+		$index = $row['ORDER_ID'];
+		$index++;
+
+		foreach ($_POST['video_sound'] as $k => $file) {
+
+			$file = explode('|@|', $file);
+
+			if ($file[0] == 'upload') {
+				$IMG_TYPE = 2;
+				$file[1] = move_video_file($file[1], 'content_' . $CID);
+			} else if ($file[0] == 'embed') {
+				$IMG_TYPE = 3;
+			} else if ($file[0] == 'link') {
+				$IMG_TYPE = 4;
+			}
+
+			unset($insert);
+			$insert['CONTENT_ID'] = $CONTENT_ID;
+			/*retrunID*/
+			$insert['IMG_TYPE'] = $IMG_TYPE;
+			$insert['IMG_PATH'] = "'" . $file[1] . "'";
+			$insert['CAT_ID'] = $CAT_ID;
+			/* cat ID*/
+			$insert['ORDER_ID'] = "'" . $index++ . "'";
+			$insert['IMG_NAME'] = "'" . $file[2] . "'";
+			$insert['DIV_NAME'] = "'" . $DIV_NAME . "'";
+			//ตั้งตาม name
+
+			$sql = "INSERT INTO trn_content_picture (" . implode(",", array_keys($insert)) . ") VALUES (" . implode(",", array_values($insert)) . ")";
+			mysql_query($sql, $conn) or die($sql);
+
+		}
+	}
+
+	if (count($_POST['video_delete_sound']) > 0) {
+		foreach ($_POST['video_delete_sound'] as $k => $file) {
+			$file = explode('|@|', $file);
+
+			if ($file[0] == 'upload') {
 				del_video_file($file[2]);
 			}
 
-			mysql_query('DELETE FROM trn_content_picture WHERE PIC_ID = '.intval($file[1]), $conn) or die($sql);
+			mysql_query('DELETE FROM trn_content_picture WHERE PIC_ID = ' . intval($file[1]), $conn) or die($sql);
 		}
 	}
-	
-	
+
+
+	//video action delete
+	if (count($_POST['video_delete_other']) > 0) {
+		foreach ($_POST['video_delete_other'] as $k => $file) {
+			$file = explode('|@|', $file);
+
+			if ($file[0] == 'upload') {
+				del_video_file($file[2]);
+			}
+
+			mysql_query('DELETE FROM trn_content_picture WHERE PIC_ID = ' . intval($file[1]), $conn) or die($sql);
+		}
+	}
+
 	if (count($_POST['video_video']) > 0) {
 		$CONTENT_ID = intval($conid);
-		$CAT_ID		= intval($CID);
-		$DIV_NAME	= 'video' ;
+		$CAT_ID = intval($CID);
+		$DIV_NAME = 'video';
 
-		$sql = "SELECT ORDER_ID FROM trn_content_picture WHERE CONTENT_ID = ".$CONTENT_ID." AND CAT_ID = ".$CAT_ID." AND DIV_NAME =  '".$DIV_NAME."' ORDER BY ORDER_ID DESC LIMIT 0 , 1";
+		$sql = "SELECT ORDER_ID FROM trn_content_picture WHERE CONTENT_ID = " . $CONTENT_ID . " AND CAT_ID = " . $CAT_ID . " AND DIV_NAME =  '" . $DIV_NAME . "' ORDER BY ORDER_ID DESC LIMIT 0 , 1";
 		$query = mysql_query($sql, $conn) or die($sql);
 		$row = mysql_fetch_array($query);
 		$index = $row['ORDER_ID'];
@@ -320,26 +414,28 @@ if (isset($_GET['edit'])) {
 
 		foreach ($_POST['video_video'] as $k => $file) {
 
-			$file = explode('|@|',$file);
+			$file = explode('|@|', $file);
 
-			if($file[0] == 'upload'){
+			if ($file[0] == 'upload') {
 				$IMG_TYPE = 2;
-				$file[1] = move_video_file($file[1],'content_' . $CID);
-			}else if($file[0] == 'embed'){
+				$file[1] = move_video_file($file[1], 'content_' . $CID);
+			} else if ($file[0] == 'embed') {
 				$IMG_TYPE = 3;
-			}else if($file[0] == 'link'){
+			} else if ($file[0] == 'link') {
 				$IMG_TYPE = 4;
 			}
 
-
 			unset($insert);
-			$insert['CONTENT_ID'] = $CONTENT_ID; /*retrunID*/
+			$insert['CONTENT_ID'] = $CONTENT_ID;
+			/*retrunID*/
 			$insert['IMG_TYPE'] = $IMG_TYPE;
 			$insert['IMG_PATH'] = "'" . $file[1] . "'";
-			$insert['CAT_ID'] = $CAT_ID; /* cat ID*/
+			$insert['CAT_ID'] = $CAT_ID;
+			/* cat ID*/
 			$insert['ORDER_ID'] = "'" . $index++ . "'";
 			$insert['IMG_NAME'] = "'" . $file[2] . "'";
-			$insert['DIV_NAME'] = "'".$DIV_NAME."'"; //ตั้งตาม name
+			$insert['DIV_NAME'] = "'" . $DIV_NAME . "'";
+			//ตั้งตาม name
 
 			$sql = "INSERT INTO trn_content_picture (" . implode(",", array_keys($insert)) . ") VALUES (" . implode(",", array_values($insert)) . ")";
 			mysql_query($sql, $conn) or die($sql);
@@ -349,18 +445,18 @@ if (isset($_GET['edit'])) {
 	//video action delete
 	if (count($_POST['video_delete_video']) > 0) {
 		foreach ($_POST['video_delete_video'] as $k => $file) {
-			$file = explode('|@|',$file);
+			$file = explode('|@|', $file);
 
-			if($file[0] == 'upload'){
+			if ($file[0] == 'upload') {
 				del_video_file($file[2]);
 			}
 
-			mysql_query('DELETE FROM trn_content_picture WHERE PIC_ID = '.intval($file[1]), $conn) or die($sql);
+			mysql_query('DELETE FROM trn_content_picture WHERE PIC_ID = ' . intval($file[1]), $conn) or die($sql);
 		}
 	}
-	
+
 	if (count($_POST['order_position']) > 0) {
-	 
+
 		foreach ($_POST['order_position'] as $k => $val) {
 			$update = "";
 			$update[] = "ORDER_ID = " . $val;
